100-days-of-algorithms/src/main/java/com/algos/searching/BinarySearch.java
package com.algos.searching;

public class BinarySearch {
    public int search(int[] nums, int num) {
        if(nums == null || nums.length == 0) {
            return -1;
        }

        int start = 0;
        int end = nums.length - 1;

        while(start <= end) {
            int mid = start + (end - start) / 2;

            if(nums[mid] == num) {
                return mid;
            } else if (num < nums[mid]) {
                end = mid - 1;
            } else {
                start = mid + 1;
            }
        }

        return -1;
    }
}
